requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
click==8.1.7
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg==3.1.10
psycopg2==2.9.7
pytz==2023.3.post1
sentry-sdk==1.30.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
